refactor: Move sink to `PennaLogEventBuilder`
This might look "out of place" initially but it makes sense:
- Both the sink and the event builder have to be unique per-thread;
- Therefore, both were wrapped in a `ThreadLocal<..>` object;
- The sink itself doesn't need to exist in the logger;
- Because the sink was in the logger, the builder had to call back the
  logger (logger -> guard -> builder -> logger), which makes this
  indirection confusing;

With this new approach, we get a tiny performance improvement by
removing one ThreadLocal, but more importantly, we made room for more
improvements in this structure as the logic is more clean and organized.
