+++ penna-core/src/main/java/penna/core/logger/guard/LevelGuard.java
@@ -1,6 +1,7 @@
 package penna.core.logger.guard;
 
 import penna.api.config.Config;
+import penna.core.logger.PennaLogEventBuilder;
 import penna.core.logger.PennaLogger;
 import org.slf4j.event.Level;
 import org.slf4j.spi.LoggingEventBuilder;
@@ -9,7 +10,7 @@ import java.util.EnumMap;
 
 /**
  * The Level guard is a property of the {@link PennaLogger} that proxies the
- * {@link penna.core.logger.event.PennaLogEventBuilder} creation.
+ * {@link PennaLogEventBuilder} creation.
  * This is done in order for us to avoid runtime checks for "is X-level allowed".
  * <br />
  * By introducing the {@link LevelGuard} as a thin proxy in the logger we allow better control over the behavior
